Feat/fix codeql (#2022)
* fix ConfigUtil.java CodeQL

* fix code checkstyle

+++ hugegraph-core/src/main/java/com/baidu/hugegraph/backend/store/BackendTable.java
@@ -193,13 +193,13 @@ public abstract class BackendTable<Session extends BackendSession, Entry> {
                 count = 1;
             }
             long maxKey = this.maxKey();
-            double each = maxKey / count;
+            Double each = maxKey / count;
 
             List<Shard> splits = new ArrayList<>((int) count);
             String last = START;
             long offset = 0L;
             while (offset < maxKey) {
-                offset += each;
+                offset += each.longValue();
                 if (offset > maxKey) {
                     splits.add(new Shard(last, END, 0L));
                     break;
